Update project files; fix Node2D and tangents
Project: Update Xcode project entries — add fs_xml_inline.{c,h} and theme_palette.h to build/headers, reinsert SendMessageAction, and rename several .xml file references to .cgen (filesystem, core, SpriteKit, DarkReign, UIKit, SceneKit) and adjust PBXGroup children ordering. Code: cast enum to int in Node2D_GetSize to ensure correct indexing (Node2D.c) and initialize vec3 n to zero in r_tangents.h to avoid uninitialized-read/UB when generating tangents. These changes fix build/project organization and address potential runtime/UB issues.

+++ plugins/UIKit/Node2D.c
@@ -213,7 +213,7 @@ HANDLER(Node2D, Node, MouseMoved)
 }
 
 float Node2D_GetSize(struct Node2D *pNode2D, enum Direction axis, enum Sizing sizing) {
-  float const size = Node2D_GetFrame(pNode2D, kBox3FieldWidth + axis);
+  float const size = Node2D_GetFrame(pNode2D, kBox3FieldWidth + (int)axis);
   switch (sizing) {
     case kSizingPlusMargin:
       return size + TOTAL_MARGIN(pNode2D, axis);
